refactor: co-locate server files with their modules
Moves single-consumer files into the modules that own them, renames the
testing fixtures directory to make its purpose obvious, and removes
noUncheckedIndexedAccess together with the defensive ceremony it forced.

- scope-resolver -> trackers/ (only used by trackers/jira.ts)
- github-client, gitlab-client -> code-hosts/
- jira-client -> trackers/
- testing/ -> __test-support__/ to signal it's test plumbing, not an app module
- drop noUncheckedIndexedAccess in both tsconfigs and remove the resulting
  /* v8 ignore */ guards, assertNever calls, and types/exhaustive.ts
