+++ server/code-hosts/__tests__/github.integration.test.ts
@@ -1,9 +1,9 @@
 import { HttpResponse, http } from "msw";
 import { describe, expect, it } from "vitest";
-import { createGitHubClient } from "../../github-client.ts";
-import { GITHUB_API, REPO } from "../../testing/support/fixtures.ts";
-import { server } from "../../testing/support/msw.ts";
+import { GITHUB_API, REPO } from "../../__test-support__/support/fixtures.ts";
+import { server } from "../../__test-support__/support/msw.ts";
 import { githubCodeHostAdapter } from "../github.ts";
+import { createGitHubClient } from "../github-client.ts";
 
 const adapter = githubCodeHostAdapter(createGitHubClient("test-token"));
