+++ server/trackers/__tests__/jira-client.integration.test.ts
@@ -1,12 +1,12 @@
 import { HttpResponse, http } from "msw";
 import { describe, expect, it } from "vitest";
-import { createJiraClient } from "../jira-client.ts";
 import {
 	createJiraIssue,
 	JIRA_API,
 	JIRA_BASE_URL,
-} from "../testing/support/fixtures.ts";
-import { server } from "../testing/support/msw.ts";
+} from "../../__test-support__/support/fixtures.ts";
+import { server } from "../../__test-support__/support/msw.ts";
+import { createJiraClient } from "../jira-client.ts";
 
 describe("Jira client", () => {
 	it("defaults search maxResults to 100", async () => {
