--- server/types/exhaustive.ts
@@ -1,3 +1,0 @@
-export function assertNever(value: never): never {
-	throw new Error(`Unexpected value: ${String(value)}`);
-}
