+++ server/workflow/__tests__/prompt-preprocessor.test.ts
@@ -1,7 +1,7 @@
 import { access, realpath } from "node:fs/promises";
 import { join } from "node:path";
 import { describe, expect, it } from "vitest";
-import { createTestWorkspaceRoot } from "../../testing/support/test-workspace.ts";
+import { createTestWorkspaceRoot } from "../../__test-support__/support/test-workspace.ts";
 import type { Issue } from "../../trackers/types.ts";
 import { issueKey, issueNumber, repoSlug } from "../../types/brands.ts";
 import {
