+++ tsconfig.json
@@ -5,7 +5,6 @@
 		"moduleResolution": "Node16",
 		"strict": true,
 		"exactOptionalPropertyTypes": true,
-		"noUncheckedIndexedAccess": true,
 		"noPropertyAccessFromIndexSignature": true,
 		"noImplicitOverride": true,
 		"noImplicitReturns": true,
